Extract styles to their own files
