+++ WooCommerce/Classes/POS/Utils/POSFontStyle.swift
@@ -1,29 +1,5 @@
 import SwiftUI
 
-/// iOS spacing style definitions for POS
-/// 1qcjzXitBHU7xPnpCOWnNM-fi-21_7403
-enum POSSpacing {
-    static let none: CGFloat = 0
-    static let xSmall: CGFloat = 4
-    static let small: CGFloat = 8
-    static let medium: CGFloat = 16
-    static let large: CGFloat = 24
-    static let xLarge: CGFloat = 32
-    static let xxLarge: CGFloat = 40
-}
-
-/// iOS padding style definitions for POS
-/// 1qcjzXitBHU7xPnpCOWnNM-fi-21_7450
-struct POSPadding {
-    static let none: CGFloat = 0
-    static let xSmall: CGFloat = 4
-    static let small: CGFloat = 8
-    static let medium: CGFloat = 16
-    static let large: CGFloat = 24
-    static let xLarge: CGFloat = 32
-    static let xxLarge: CGFloat = 40
-}
-
 /// iOS type style definitions for POS
 /// 1qcjzXitBHU7xPnpCOWnNM-fi-23_7310
 enum POSFontStyle {
