+++ WooCommerce/Classes/POS/Utils/POSPadding.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+/// iOS padding style definitions for POS
+/// 1qcjzXitBHU7xPnpCOWnNM-fi-21_7450
+struct POSPadding {
+    static let none: CGFloat = 0
+    static let xSmall: CGFloat = 4
+    static let small: CGFloat = 8
+    static let medium: CGFloat = 16
+    static let large: CGFloat = 24
+    static let xLarge: CGFloat = 32
+    static let xxLarge: CGFloat = 40
+}
